% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-reeb-pairs.R
\name{as.data.frame.reeb_graph_pairs}
\alias{as.data.frame.reeb_graph_pairs}
\title{An S3 class for Reeb graph persistent pairs}
\usage{
\method{as.data.frame}{reeb_graph_pairs}(x, ...)
}
\description{
This S3 class is a light wrapper around a data frame containing
the values, indices, and orders of persistent pairs.
}
\details{
The data frame has 6 columns, 2 each (birth and death) for 6
properties: value (height), index, and order.
}
\examples{
x <- reeb_graph(
  values = c(0, .4, .6, 1),
  edgelist = c( 1,2, 1,3, 2,4, 3,4 )
)
mp <- reeb_graph_pairs(x)
print(mp)
as.data.frame(mp)
}
