% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{glomeruli}
\alias{glomeruli}
\alias{Nyengaard1993}
\title{Glomerular Capillary Networks}
\format{
2-column matrices encoding undirected graphs indexed from 1:
\describe{
\item{\code{Nyengaard1993}}{newborn rat (Nyengaard & Marcussen, 1993)}
}
}
\usage{
Nyengaard1993
}
\description{
These edge lists encode the capillary structure of several
murine glomeruli.
}
\details{
As meta-analyzed by Wahl &al (2004), several investigations of the
capillary systems in (mostly rat) glomeruli have produced transit-style
diagrams of their network structure. Some of these painstakingly
reconstructed networks are used to illustrate data analysis using Reeb
graphs.

The transit diagrams were encoded as edgelists via breadth-first search. In
each case, vertex \code{1} is the afferent arteriole and the highest-index
vertex is the efferent arteriole.

There is no intrinsic height function. Instead, circuit analysis can be
used to infer direction of flow and assign pressure differentials to the
vertices.
}
\references{
Nyengaard JR & Marcussen N (1993) "The number of glomerular capillaries
estimated by an unbiased and efficient stereological method". \emph{Journal of
Microscopy} 171(1): 27--37. \doi{10.1111/j.1365-2818.1993.tb03356.x}

Wahl EM, Quintas LV, Lurie LL, & Gargano ML (2004) "A graph theory analysis
of renal glomerular microvascular networks". \emph{Microvascular Research} 67(3):
223--230. \doi{10.1016/j.mvr.2003.11.005}
}
\keyword{datasets}
