% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{reeb_graph_examples}
\alias{reeb_graph_examples}
\alias{buddha}
\alias{david}
\alias{topology}
\alias{flower}
\title{Mesh-Derived Reeb Graphs}
\format{
Objects of class \code{reeb_graph}.
}
\source{
http://visionair.ge.imati.cnr.it/ontologies/shapes/ (\code{AIM@SHAPE}; defunct)
\url{https://github.com/USFDataVisualization/ReebGraphPairing}
}
\usage{
david
buddha
topology
flower
}
\description{
These are some of the Reeb graphs used by Tu &al (2019; Fig. 8)
to benchmark the multi-pass and single-pass algorithms. The original meshes
were obtained from the \code{AIM@SHAPE} Shape Repository, while the authors
computed Reeb graphs using a custom C++ implementation.
}
\keyword{datasets}
