% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgp-package.R
\docType{package}
\name{rgp}
\alias{rgp-package}
\alias{rgp}
\title{rgp: Compute Extended Persistent Homology of Reeb Graphs}
\description{
Interface to the ReebGraphPairing Java program \doi{10.1007/978-3-030-33720-9_8} via rJava.
}
\author{
\strong{Maintainer}: Satyajit Mohanty \email{mohantysatyajit57@gmail.com}

Authors:
\itemize{
  \item Shubham Singh \email{shubhamus16091998@example.com}
  \item Jason Cory Brunson \email{cornelioid@gmail.com} (\href{https://orcid.org/0000-0003-3126-9494}{ORCID}) [laboratory director]
}

Other contributors:
\itemize{
  \item Paul Rosen \email{paul.rosen@utah.edu} (\href{https://orcid.org/0000-0002-0873-9518}{ORCID}) (ReebGraphPairing; GPL-3 license) [copyright holder, contributor]
  \item Junyi Tu \email{JXTU@salisbury.edu} (\href{https://orcid.org/0000-0001-7026-7454}{ORCID}) (ReebGraphPairing; GPL-3 license) [copyright holder, contributor]
}

}
\keyword{internal}
