% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgph-package.R
\docType{package}
\name{rgph}
\alias{rgph-package}
\alias{rgph}
\title{rgph: Pair Critical Points and Compute Persistent Homology of Reeb Graphs}
\description{
Interface to the 'ReebGraphPairing' program to compute critical points of Reeb graphs following Tu, Hajij, & Rosen (2019) \doi{10.1007/978-3-030-33720-9_8} via the 'rJava' package. Also store Reeb graphs in a minimal S3 class, convert between other network data structures, and post-process pairing data to obtain extended persistent homology following Carrière & Oudot (2018) \doi{10.1007/s10208-017-9370-z}.
}
\author{
\strong{Maintainer}: Jason Cory Brunson \email{cornelioid@gmail.com} (\href{https://orcid.org/0000-0003-3126-9494}{ORCID})

Authors:
\itemize{
  \item Satyajit Mohanty \email{mohantysatyajit57@gmail.com}
  \item Shubham Singh \email{shubhamus16091998@example.com}
}

Other contributors:
\itemize{
  \item Paul Rosen \email{paul.rosen@utah.edu} (\href{https://orcid.org/0000-0002-0873-9518}{ORCID}) (ReebGraphPairing; GPL-3 license) [copyright holder, contributor]
  \item Junyi Tu \email{JXTU@salisbury.edu} (\href{https://orcid.org/0000-0001-7026-7454}{ORCID}) (ReebGraphPairing; GPL-3 license) [copyright holder, contributor]
}

}
\keyword{internal}
