% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reeb-graph-class.R
\name{reeb_graph}
\alias{reeb_graph}
\alias{print.reeb_graph}
\alias{format.reeb_graph}
\alias{read_reeb_graph}
\title{An S3 class and constructors for Reeb graphs}
\usage{
reeb_graph(values, edgelist)

\method{print}{reeb_graph}(x, ..., n = NULL)

\method{format}{reeb_graph}(x, ..., n = NULL)

read_reeb_graph(file)
}
\arguments{
\item{values}{Numeric vector of function values at vertices.}

\item{edgelist}{2-column integer matrix of linked vertex pairs.}

\item{x}{Object of class \code{"reeb_graph"}.}

\item{...}{Additional arguments passed to \code{\link[base:format]{base::format()}}.}

\item{n}{Integer number of edges to print.}
}
\value{
An object of class \code{"reeb_graph"}, which is a list of two elements:
\itemize{
\item \code{values}: Numeric vector of function values at vertices.
\item \code{edgelist}: 2-column integer matrix of linked vertex pairs.
}
}
\description{
This is an S3 class with associated constructors for a data structure to
represent Reeb graphs in R.
}
\details{
Vertex indices start at zero, for consistency with examples. The
positions of \code{values} and the integer values in \code{edgelist} will correspond
to the same vertices; \code{length(values) - 1L} must bound \code{max(edgelist)}.
}
\examples{
x <- reeb_graph(
  values = c(0, .4, .6, 1),
  edgelist = rbind(c(0, 1), c(0, 2), c(1, 3), c(2, 3))
)
print(x)

read_reeb_graph("files/running_example_reeb_graph.txt")

}
