% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reeb-graph-pairs.R
\name{reeb_graph_pairs}
\alias{reeb_graph_pairs}
\title{Pair Reeb Graph Critical Points via Java}
\usage{
reeb_graph_pairs(x, method = "multi_pass")
}
\arguments{
\item{x}{A \code{\link{reeb_graph}} object.}

\item{method}{Character; the pairing method to use. Matched to
\code{"single_pass"} or \code{"multi_pass"}.}
}
\value{
A data frame containing the six output vectors returned by the Java
method: the birth and death values (\code{double}), birth and death indices
(\code{integer}), and birth and death orders (\code{integer}). The data frame has
attributes \code{"method"} for the method used and \code{"elapsedTime"} for the
elapsed time.
}
\description{
This function calls one of two methods, merge-pair and
propagate-and-pair, to pair the critical points of a Reeb graph.
}
\details{
The function uses the \code{rJava} package to call either of two Java
methods from \code{ReebGraphPairing}. Ensure the Java Virtual Machine (JVM) is
initialized and the required class is available in the class path.
}
\examples{
ex_sf <- system.file("extdata", "running_example.txt", package = "rgp")
( ex_rg <- read_reeb_graph(ex_sf) )
( ex_cp <- reeb_graph_pairs(ex_rg) )
attr(ex_cp, "method")
attr(ex_cp, "elapsedTime")
}
