% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reeb-graph-persistence.R
\name{reeb_graph_persistence}
\alias{reeb_graph_persistence}
\title{Compute Persistent Homology of a Reeb Graph}
\usage{
reeb_graph_persistence(
  x,
  sublevel = TRUE,
  method = c("single_pass", "multi_pass"),
  value = c("value", "index", "order")
)
}
\arguments{
\item{x}{A \code{\link{reeb_graph}} object.}

\item{sublevel}{Logical; whether to take the sublevel set filtration (\code{TRUE},
the default) or else the superlevel set filtration (via reversing
\code{x$values} before paring critical points.}

\item{method}{Character; the pairing method to use. Matched to
\code{"single_pass"} (the default) or \code{"multi_pass"}.}

\item{value}{Character; the numerical value used by the persistent pairs.
Matched to \code{"value"} (the default), \code{"index"}, or \code{"order"}.}
}
\value{
A \link[phutil:persistence]{phutil::persistence} object.
}
\description{
This function obtains extended persistent homology of a Reeb
graph by way of pairing critical points.
}
\details{
The types, values, and indices of critical pairs are obtained by
\code{\link[=reeb_graph_pairs]{reeb_graph_pairs()}}. \code{reeb_graph_persistence()} calls this function
internally with the prescribed \code{method}, then restructures the values or
indices as \link[phutil:persistence]{phutil::persistence} data.

This function may be deprecated once a \code{"reeb_graph_pairs"} method is
written for \code{\link[phutil:persistence]{phutil::as_persistence()}}.
}
\examples{
ex_sf <- system.file("extdata", "running_example.txt", package = "rgp")
( ex_rg <- read_reeb_graph(ex_sf) )
( ex_ph <- reeb_graph_persistence(ex_rg) )
phutil::get_pairs(ex_ph, dimension = 0)
phutil::get_pairs(ex_ph, dimension = 1)

t10_f <- system.file("extdata", "10_tree_iterations.txt", package = "rgp")
( t10 <- read_reeb_graph(t10_f) )
( t10_ph <- reeb_graph_persistence(t10) )
phutil::get_pairs(t10_ph, dimension = 0)
( t10_ph <- reeb_graph_persistence(t10, value = "index") )
phutil::get_pairs(t10_ph, dimension = 0)
( t10_ph <- reeb_graph_persistence(t10, value = "order") )
phutil::get_pairs(t10_ph, dimension = 0)

}
